Fix tracking number not being propagated to order cartons
`Spree::OrderShipping#ship` already copies the tracking number from the
shipment to the carton, as long as the tracking number is present on
the shipment when method is called.

`ShipmentNotice#ship_shipment`, however, was only setting the tracking
number *after* calling `Spree::Shipment#ship!`, so the tracking number
was not being set on the shipment's carton.