Handle double shipments gracefully
When a label is voided in ShipStation and a new one is generated, or
when an order is re-shipped in ShipStation for any reason, ShipStation
will send another webhook with the same order number and the new tracking
information.

Before this change, the `ShipmentNotice` service object would attempt to
call `#ship!` on the corresponding shipment every time it received a
webhook from ShipStation. A shipped shipment cannot be re-shipped in
Solidus, so this would cause an error.

With the change, subsequent webhooks for the same shipment will skip the
`#ship!` call and will just update the tracking number (and run the
payment processing logic, if needed).

Note that this change assumes that a double shipment in ShipStation means
that a label was voided and re-generated for an order, but it could also
mean that multiple labels were sent for a given order (e.g., if an order
was split into multiple shipments when in ShipStation). This would be
pretty complicated to infer, so for the time being we assume it's just
a label re-generation.