Reduce repetition in ShipmentNotice spec
I'm definitely not a fan of mystery guests introduced by `let` and
`before`, but in this case the amount of repetition and nested context
was making the spec was way too complicated to follow.

This is probably a sign that `ShipmentNotice` does too many things, but
until we have time for a proper refactor, the spec change should at least
make the test a bit easier to understand.