Stop checking the delta between last sync and last update
Previously, we would discard shipments were the last sync happened too
far in the past compared to the last update. The idea here was to discard
bad shipments and stop trying to sync them over and over again, but the
result was that, if you change an order more than e.g. 7 days after its
last ShipStation sync (7 days being the threshold), it will not sync anymore.

The new solution is to always compare the threshold against the last
order update: if the order has been updated more than 7 days ago and *still*
needs a sync, then we don't sync it, because we assume it's invalid.

This way, it's also possible to trigger a ShipStation re-sync by simply
recalculating the order.