Handle invalid shipment batches gracefully
When an entire batch of shipments is invalid, `#bulk_create_orders`
returns early, which means the response will be empty. This would
cause an error when accessing the `results` hash key.

Since the error is already being handled by `#bulk_create_orders`
when attempting to serialize the shipments, we don't need to crash
the batch syncer too, and we can just handle the situation gracefully.